% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dividir.R
\name{dividir}
\alias{dividir}
\title{Função para dividir}
\usage{
dividir(a, b)
}
\arguments{
\item{a}{um numero}

\item{b}{outro numero}
}
\description{
Esta função será utilizada para dividir dois numeros
}
\examples{
Dividir(8, 4)


}
