% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplicar.R
\name{multiplicar}
\alias{multiplicar}
\title{Função para multiplicar}
\usage{
multiplicar(a, b)
}
\arguments{
\item{a}{um numero}

\item{b}{outro numero}
}
\description{
Esta função será utilizada para multiplicar dois numeros
}
\examples{
multiplicar(3, 4)


}
