% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/somar.R
\name{somar}
\alias{somar}
\title{Função para somar}
\usage{
somar(a, b)
}
\arguments{
\item{a}{um numero}

\item{b}{outro numero}
}
\description{
Esta função será utilizada para somar dois numeros
}
\examples{
somar(3, 4)


}
