% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtrair.R
\name{subtrair}
\alias{subtrair}
\title{Função para subtrair}
\usage{
subtrair(a, b)
}
\arguments{
\item{a}{um numero}

\item{b}{outro numero}
}
\description{
Esta função será utilizada para subtrair dois numeros
}
\examples{
subtrair(3, 4)


}
